requirements.txt
cmake==3.12.0
Cython==0.28.5
dlib==19.15.0
face-recognition==1.2.3
face-recognition-models==0.3.0
h5py==2.8.0
imutils==0.4.6
numpy==1.15.1
opencv-contrib-python==3.4.2.17
opencv-python==3.4.2.17
Pillow==5.2.0
simplejson==3.16.0
tensorboard==1.9.0
tensorflow==1.9.0
Werkzeug==0.14.1
